+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     double click on files to prompt for open with UVtools, it also allow to select it as default program (maybe fix
     #459)
 - (Add) Dynamic lifts: Configure separate bottom/normal retract speeds and two-stage retract distance percentages (#927)
+- (Add) Configure bottom height and transition layers for wait time after cure (#1093)
 - (Fix) Align pixel editor line brush preview and applied position (#966)
 - (Improvement) pixel editor drawing with smooth, interpolated strokes committed on pointer release by @jorgerobles
   (#1139)